+++ packages/core/src/index.ts
@@ -39,6 +39,7 @@ export * from "./hooks/UseMouseIsEntered";
 export * from "./hooks/UseMultiOnClickOutside";
 export * from "./hooks/UseOnClickOutside";
 export * from "./hooks/UseOnNoMouseInput";
+export * from "./hooks/UseForwardedRef";
 
 export * from "./utils/SwitchCaseExhauster";
 export * from "./utils/parsers/NumberParser";