+++ CHANGELOG.md
@@ -33,6 +33,10 @@ Example:
 
 - `NavBarButton` icon now has same color as the text.
 
+### ButtonGroup
+
+- `ButtonGroup` no longer has outer margin.
+
 ### reducerIdGate
 
 Add helper function for wrapping actions with reducerIdGate action.
@@ -53,6 +57,7 @@ It returns the same actions, but wrapped with reducerIdGate action.
 - `ActionMenuLink` can be triggered with Space, just like `ActionMenuItem`
 - `ActionMenuLink` has border radius if first or last child, just like `ActionMenuItem`
 - The text in `ActionMenuItem` and `ActionMenuLink` does not wrap
+- New component `ActionMenuLinkContent` which can be used to display an item that is not a button.
 
 ### StandardTable